requirements.txt
jupyterlab
pandas
matplotlib
jupyter-lsp
jupyterlab-lsp
python-lsp-server[all]
Pillow
scipy
